Separated GridClass andCellClass into new files, also create two different functions for creating a grid overlay. Discovered that one of them is on average a bit faster and made it the default
