+++ vidtomu.py
@@ -1,68 +1,14 @@
 import numpy as np
 import cv2
+from GridClass import GridFrame
 
 cap = cv2.VideoCapture(0)
 
 #   I think I will need to create a custom class for my purposes
 #   as well as a grid listener
 
-#   Grid Frame Class that will serve as base for my project
-#   It'll carry the frame and the Grid Listener
-class GridFrame:
-    def __init__(self, frame, cellSize):
-        self.frame = frame
-        self.cellSize = cellSize
-        self.fHeight, self.fWidth, self.fChannels = frame.shape
-
-        #   Array that later will hold the cells of the grid
-        self.GridCells = []
-        self.createGrid()
-
-    #   Shows the frame that is held by the GridFrame
-    def showGridFrame(self):
-        cv2.imshow('frame',self.frame)
-
-    #   Attempts to show A cell of the grid
-    def showCell(self, cellNumber):
-        self.GridCells[cellNumber].showC()
-
-    #   Show all cells
-    def showCells(self):
-        for c in self.GridCells:
-            c.showC()
-
-    #   Also populates the GridCells Array that is a member of this class
-    #   According to the cellSize specified on init
-    def createGrid(self):
-        cellHeight = int(self.fHeight/self.cellSize)
-        cellWidth = int(self.fWidth/self.cellSize)
-        for x in range(0,self.cellSize):
-            for y in range(0, self.cellSize):
-                # Create a new GridCell Instance and initialize
-                cell = GridCell(cellHeight, cellWidth, cellHeight*y, cellWidth*x,
-                                self.frame[cellHeight*y:cellHeight*y+cellHeight,cellWidth*x:cellWidth*x+cellWidth])
-                self.GridCells.append(cell)
-
-# Meta Object that will describe a singular cell in the GridFrame
-# Will be used to find events occuring inside these cells
-class GridCell():
-    def __init__(self, height, width, startX, startY, partialFrame):
-        self.height = height
-        self.width = width
-        self.startX = startX
-        self.startY = startY
-        self.partialFrame = partialFrame
-
-    def showC(self):
-        cv2.imshow('cell',self.partialFrame)
-
-    # should create an overlay with an translucent color
-    def drawBorders(self):
-        # Color the border of each cell here
-        x = 0
-
-
-#Divides Image or frame into several images depending on subdivision rate
+### This is now Obsolete ### MIght be salvaged for other bits of Code ###
+#   Divides Image or frame into several images depending on subdivision rate
 def subdivide(frame, numOfSubFrms):
     #cloning the original frame such that we'll be working on a copy from now on
     frameC = np.array(frame, copy = True)
@@ -112,6 +58,24 @@ def colorMask(frame, lower, upper):
     res = cv2.bitwise_and(frame, frame, mask = mask)
     return frame, mask, res
 
+#   Method to time average run time of functions that are passed
+#   Dunno how to use this yet lmao
+#   average and counter should be passed as 0 and 1 respectively
+def measureTime(average, counter, methodToTime):
+    import time
+    start = time.time()
+
+    methodToTime()
+
+    end = time.time()
+    print(end - start)
+    print("Execution time above ^^^")
+    print("")
+    time = end - start
+    average += time
+    counter += 1
+    print("Average Time: ", average/counter, "Counter: ", counter)
+
 
 # here we display the original frame alongside the mask that we created and the colorshifted picture
 def show(frame, mask, res):
@@ -142,9 +106,15 @@ while(1):
 
     #subdivide(frame, 2)
 
-    GF = GridFrame(frame,2)
+    GF = GridFrame(frame,3)
+
     GF.showGridFrame()
+
     GF.showCells()
+
+    GF.drawCellBorders()
+
+    GF.drawGrid()
 
     k = cv2.waitKey(5) & 0xFF
     if k == 27:
